+++ FabControl/FloatingActionButtonSize.cs
@@ -1,9 +1,18 @@
 ﻿
 namespace Refractored.FabControl
 {
+    /// <summary>
+    /// Size of FAB
+    /// </summary>
     public enum FloatingActionButtonSize
     {
+        /// <summary>
+        /// Normal size
+        /// </summary>
         Normal,
+        /// <summary>
+        /// Mini size
+        /// </summary>
         Mini
     }
 }